updates comment format
